.webgen/processor/pyproject.toml
[project]
name = "processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.52.2",
    "python-dotenv>=1.0.1",
]
